+++ h3ronpy/h3ronpy/util.py
@@ -1,9 +1,8 @@
-import typing
-
 import geopandas as gpd
 import h3.api.numpy_int as h3
 import numpy as np
 import pandas as pd
+import typing
 from shapely.geometry import Polygon
 
 from . import H3_CRS
@@ -16,9 +15,9 @@ def h3indexes_to_geodataframe(
     """
     i_arr = np.asarray(h3indexes)
     return gpd.GeoDataFrame({
-        "geometry": [Polygon(h3.h3_to_geo_boundary(h, geo_json=True)) for h in np.nditer(i_arr)],
-        "h3resolution": [h3.h3_get_resolution(h) for h in np.nditer(i_arr)],
-        "h3index": [h3.h3_to_string(h) for h in np.nditer(i_arr)]
+        "geometry": [Polygon(h3.h3_to_geo_boundary(h, geo_json=True)) for h in _array_nditer(i_arr)],
+        "h3resolution": [h3.h3_get_resolution(h) for h in _array_nditer(i_arr)],
+        "h3index": [h3.h3_to_string(h) for h in _array_nditer(i_arr)]
     }, crs=H3_CRS)
 
 
@@ -31,5 +30,10 @@ def h3index_column_to_geodataframe(df: pd.DataFrame, column_name: str = "h3index
     :return: GeoDataFrame
     """
     return gpd.GeoDataFrame(df,
-                            geometry=[Polygon(h3.h3_to_geo_boundary(h, geo_json=True)) for h in np.nditer(df[column_name].to_numpy())],
+                            geometry=[Polygon(h3.h3_to_geo_boundary(h, geo_json=True)) for h in
+                                      _array_nditer(df[column_name].to_numpy())],
                             crs=H3_CRS)
+
+
+def _array_nditer(a: np.array):
+    return np.nditer(a, flags=["zerosize_ok", ])